+++ docs/recipes/extend/config.js
@@ -60,7 +60,7 @@ appe__config = {
   "runtime_path": "app",
   "save_path": "save",
   "aux_path": "ext",
-  "open_attempts": 20,
+  "load_attempts": 100,
   "alt": {
     "exec_folder": "alt",
     "exec_platform": {